refactor(#186): SIP-0097 slice 4 — pulse path moves to PulseBoundaryRunner
Extract the six pulse methods verbatim into
adapters/cycles/pulse_boundary_runner.py (SIP-0097 §6.2): the executor's
three call sites (setup_pulse_context / evaluate_pulse_boundaries /
run_pulse_evaluations) plus the internal _verify_with_repair /
_run_boundary_verification / _emit_pulse_event. Verification summaries
keep appending to the RunLedger threaded from execute_run (§6.6);
exhaustion still raises the same _ExecutionError the run loop converts
to FAILED.

Same interim seam as slice 3 (§6.2/AC#9): three late-bound
executor-supplied callables — dispatch_task + create_task_run
(re-pointed to TaskDispatcher in slice 5) and store_artifact (§6.7
executor residual). LLM observability is a real constructor dep but
event-emission-only per the observability ownership rule (no scope
open/close). No cancellation probe: the pulse path never checked
cancellation; it relies on dispatch-boundary checks per the §6 rule.

Mechanical test updates (disclosed per §7.7): run_pulse_verification
patch targets re-pointed to the new module (25 sites); the cadence
time-simulation test patches time.monotonic in BOTH modules (executor
resets cadence_start_time, runner computes elapsed); the pulse-repair
propagation test re-pointed to the collaborator. Emission guardrails
re-anchored: executor 17→12 EventType refs, PulseBoundaryRunner owns 8
(5 PULSE_* + repair-chain TASK_*); total emit sites unchanged at 41.

The moved methods pass C901=12 at their new home — no new per-file
ignore; pyproject.toml untouched. Executor: 2,632 → 2,036 lines.
